Balance at the Transport Level.
Previous implementation was dump. This creates a new object `Transport`
that implements `http.RoundTripper` to do the balancing. This creates
one choke point for updating the host and is a lot cleaner.

Transport still rewrites the host and passes it to the
http.DefaultTransport by default. This transport does its own
connection pooling by host:port.

Alternately one could do the balancing at the TCP dial level but this
would mess with the http package’s internal balancing allowing its idle
pool to potentially contain connections to just one of the hosts.