Handle missing node on watch event.
When things change, a watch will be reset on the parent node and its
children. This process returns the set of child keys. Right after,
those keys are check for their values sequentially, to update the
endpoints list. It is possible, while the children are being checked,
one of them disappears. This case is now handled properly.